Hotfix: refactor concourse pipeline to match dev docs
Currently we're dealing with potential secrets leakage management...
The way we're deploying things is printing out the credentials to the
pipeline. This isn't the worst thing ever, as Concourse is only
accessible by people within the Bardeen network.

This however, is still potential threat... It could lead to GovWifi
product page being overwritten with no audit of when, why or who...

This makes the pipeline at least consistent with other pipelines and removes dependency on Dockerfile and Makefile

This also removes the need to create a deploy a docker container to ecr

Signed-off-by: Paula Valenca <paula.valenca@digital.cabinet-office.gov.uk>