Hotfix: copy S3 files with bundled before deployment
The pipeline was retrieving the S3 files but was not bundling
them before deployment.
Because there are default files in place, the build does not fail
and the test files are deployed instead.

The reason the files were being copied before was because this
line on the Dockerfile was masking the bug:
https://github.com/alphagov/govwifi-product-page/blob/b51e52575f7b6606a8ce048242b94d8470d673dd/Dockerfile#L11

This is similar to the issue we saw before on 49bc9c49b434f995a6123771a0845776db340628

Signed-off-by: Paula Valenca <paula.valenca@digital.cabinet-office.gov.uk>