Pin tag docker image instead of latest
Instead of always using the latest version of the base docker image,
pin to a specific digest..

We are using the latest at the time of this commit, which corresponds
to tag 7a7678281e1f152183c7359546912df68f298664.
See https://hub.docker.com/r/governmentpaas/cf-cli/tags

See https://github.com/concourse/concourse/issues/2439 regarding resource pinning
on Concourse.

Also note that the pipeline will not check for the resource if the pinned
version already exists locally, so you may not see which digest is being
used on the job logs on concource web ui. See for context,
https://github.com/concourse/concourse/issues/2443

Signed-off-by: Paula Valenca <paula.valenca@digital.cabinet-office.gov.uk>