Tell admins how to raise support tickets when not signed in
If we’re worried that admin users will get stuck trying to find where to
go for support when not signed in, now that we’ve removed the link to
the support form from the sign in page[1], we should change this.

I’ve tried to word the link in such a way that end-users won’t think
this is the place for them to get support.

1. https://github.com/alphagov/govwifi-admin/pull/824