css: correctly point to the new GovUK frontend assets
They `assets/` folder in the `govuk-frontend` package needs to be
pointed at with the `link_tree` directive (which has to be relative
hence the manual fetch into node_modules) so that Sprockets can
copy and deliver the assets within it.

The SCSS variable `$govuk-assets-path` is then overriden to make sure
the subsequent request points to fonts/xxx rather than
assets/fonts/xxx, as we override the default Middleman output
folder (which is `assets`).

This fixes both images and fonts being missing on the rewrite.