gem: move to libsassc
The `sass` ruby gem is deprecated and `libsassc` (gem name: `sassc`)
is a much-faster, drop-in replacement.

Unfortunately `middleman-sprockets` isn't always happy with the new
gem[1] so include the old one as well in the bundle.

As per the code changes, point towards a standard and IE-specific
manifests, which in turn include the GOV.UK Frontend assets with the
correct ie-specific variables set in the latter case.

(re)compile-time greatly reduced.

[1]: https://github.com/middleman/middleman-sprockets/issues/132