nav: fix "Connect a device"'s link and amend helper for active class
We don't always point to the root of a section in the top nav:
"Connect a device" points to `/about-govwifi/connect-to-govwifi/`
rather than just `/about-govwifi`. To make sure we can point to a
nested level of the section and preserve our `start_with?` trick,
introduce an optional `root_scope` to ensure this can be infered.