a11y: don't load two identical images on the homepage (1.3.3)
We served the homepage image twice because of our different layouts on
mobile and desktop: on mobile the image is first, text below; on
desktop, text is first and image on the left.

Use a bit of CSS to allow both versions to cohabitate with a single
markup rather than loading both and conditionally hiding either one.