fix: remove middleman-sprockets and govuk_template
middleman-sprockets is deprecated[1] so remove all of our terrible
hacks and stick with SASS import statements. Remove the hard
dependency on SASS as well since Middleman ships with it. Remove the
govuk_template dependency as we don't use it either.

This will also make the bundle smaller and faster (the node_modules
folder is +8M smaller).

[1]: https://github.com/middleman/middleman-sprockets/issues/132#issuecomment-538590940