feat: allow amending cookie preferences
This really just involves upgrading `govwifi-shared-frontend` to 0.6.4
which exposes `setCategoryAllowed(category, isAllowed)` and using it
in conjuntion with `getCategoryAllowed(category)` to setup the form
inputs.

Once the form is submitted, call the API to record the changes and
show a banner on the page to inform the user their preferences have
been saved (styling inspired by Notify).