Ensure the tests work
The tests previously ran under the development environment. The development
environment includes livereloading, which requires
loading a number of gems. This is fine when running Middleman, but
not when running tests as the spec_helper does not require the gems.

This commit sets the environment the tests run under as 'build', which
ensures the tests work.